requirements.txt
flask
flask_cors
requests
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
jwt
oauth2client
oauthlib
pymongo
requests
requests-oauthlib
simplejson